src/core/models/user.py
from django.db import models
from django.contrib.auth.base_user import AbstractBaseUser
from django.contrib.auth.models import PermissionsMixin

class User(AbstractBaseUser, PermissionsMixin):
    username = models.CharField('username', unique=True),
    password = models.CharField('password')