src/pyproject.toml
[tool.poetry]
name = "python-api"
version = "0.1.0"
description = ""
authors = ["e.shtab <shtab.eg@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2 = "^2.9.9"
dj-database-url = "^2.2.0"
django-rest-swagger = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
